% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgcv_maps_wald.R
\name{mgcv_maps_wald}
\alias{mgcv_maps_wald}
\title{Perform Wald Tests on mgcv Smooth Terms (Null and Penalized Components)}
\usage{
mgcv_maps_wald(fit, test.component = 1, null.tol = 1e-10)
}
\arguments{
\item{fit}{A fitted `gam` or `bam` model from the `mgcv` package.}

\item{test.component}{The index of smooth component to be tested.}

\item{null.tol}{The tolerance of row norm to detect indices of null space. Default to 1e-10.}
}
\value{
A `data.table` summarizing Wald statistics and p-values separately
        for null (unpenalized) and penalized (smooth) components of each term.
}
\description{
This function takes a fitted mgcv `gam` or `bam` object,
then performs separate Wald tests on the null (unpenalized/fixed) and penalized
(smooth) components of each smooth term.
It returns a clear summary as a `data.table`.
}
\examples{
\dontrun{
fit <- gam(y ~ s(x0) + s(x1), data=dat, method="REML")
mgcv_maps_wald(fit)
}

}
