% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecewise_linearity.R
\name{piecewise_linearity}
\alias{piecewise_linearity}
\title{Generate a piecewise linear basis matrix with specified changepoints}
\usage{
piecewise_linearity(x, para)
}
\arguments{
\item{x}{A numeric vector representing the covariate values at which to evaluate the piecewise linear function.}

\item{para}{A numeric scalar or numeric vector indicating the location(s) of the changepoint(s) (knot positions).}
}
\value{
A numeric matrix with columns corresponding to the intercept, the original covariate `x`, and additional columns representing the positive parts of `(x - knot)` for each specified changepoint.
}
\description{
This function creates a design matrix representing a piecewise linear function with one or multiple changepoints.
Each changepoint introduces a knot at which the slope of the function may change.
}
\examples{
x <- seq(0, 10, length.out=100)
piecewise_linearity(x, para=c(3,7))

}
