% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict.matrix.Fixed.smooth.R
\name{Predict.matrix.Fixed.smooth}
\alias{Predict.matrix.Fixed.smooth}
\title{Prediction Matrix for Fixed-Effect-Only Smooth Term}
\usage{
\method{Predict.matrix}{Fixed.smooth}(object, data)
}
\arguments{
\item{object}{A smooth term object of class `"Fixed.smooth"`, containing the stored `getA` function and `para`.}

\item{data}{A data frame containing the covariate for prediction.}
}
\value{
A matrix representing the fixed effect basis for the new data.
}
\description{
This function generates the prediction matrix for an `Fixed.smooth` object, using the stored
`getA` function and its parameters (`para`) to construct the fixed effect basis for new data.
}
