% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharp_linearity_discontinuity.R
\name{sharp_linearity_discontinuity}
\alias{sharp_linearity_discontinuity}
\title{Generate basis functions for Sharp Regression Discontinuity Design with multiple breakpoints}
\usage{
sharp_linearity_discontinuity(x, para)
}
\arguments{
\item{x}{A numeric vector representing the covariate values.}

\item{para}{A numeric scalar or vector indicating the location(s) of the changepoint(s).}
}
\value{
A numeric matrix where columns represent the intercept, the covariate `x`, and
        jump indicators `I(x > para[i])`.
}
\description{
This function creates a design matrix for a shape regression discontinuity model with multiple cutoff points.
Each cutoff introduces a shape discontinuity and a slope change at the corresponding threshold.
}
\examples{
x <- seq(0, 10, length.out=100)
sharp_linearity_discontinuity(x, para=c(3, 7))

}
