% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothed_piecewise_linearity.R
\name{smoothed_piecewise_linearity}
\alias{smoothed_piecewise_linearity}
\title{Generate a smoothed piecewise linear basis matrix with specified changepoints}
\usage{
smoothed_piecewise_linearity(x, para)
}
\arguments{
\item{x}{A numeric vector representing the covariate values at which to evaluate the smoothed piecewise linear function.}

\item{para}{A list containing:
\itemize{
  \item `para[[1]]`: A numeric vector indicating the location(s) of the changepoint(s) (knot positions).
  \item `para[[2]]`: A numeric scalar representing the smoothing parameter `b`, which controls the transition smoothness around the changepoints.
}}
}
\value{
A numeric matrix with columns corresponding to the intercept, the original covariate `x`, and additional columns representing the smoothed version of `(x - knot)_+` for each specified changepoint.
}
\description{
This function creates a design matrix representing a piecewise linear function with one or multiple changepoints,
incorporating a smooth approximation near each changepoint using a quadratic smoothing function.
}
\details{
This function extends the standard piecewise linear basis by introducing a **quadratic smoothing** around each changepoint.
The smoothing is achieved by approximating `(x - v)_+` with a quadratic spline in the interval `[v - b, ν + b]`:
\deqn{s(x, v, b) = \frac{(x - v + b)^2}{4b} 1_{[v - b, ν + b]} + (x - v) 1_{[x \geq v + b]}}
where `b` is the user-specified smoothing parameter.
}
\examples{
x <- seq(0, 10, length.out=100)
smoothed_piecewise_linearity(x, para=list(c(3,7), 0.5))

}
