% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taps_score_test.R
\name{taps_score_test}
\alias{taps_score_test}
\title{Perform Score Tests on mgcv Smooth Terms (Null and Penalized Components)}
\usage{
taps_score_test(
  fit,
  test.component = 1,
  null.tol = 1e-10,
  method = "davies",
  max_eps = 1e-06,
  max_iter = 1e+05
)
}
\arguments{
\item{fit}{A fitted `gam` or `bam` model from the `mgcv` package.}

\item{test.component}{The index of smooth component to be tested.}

\item{null.tol}{The tolerance of row norm to detect indices of null space. Default to 1e-10.}

\item{method}{Method for p-value calculation: `"satterthwaite"` or `"liu"` or `"davies"` (default) or `"hbe"` or `"wood"`.}

\item{max_eps}{The tolerance used in CompQuadForm to compute the p-values. Default to 1e-6.}

\item{max_iter}{The maximum iteration used in CompQuadForm to compute the p-values. Default to 1e5.}
}
\value{
A `data.table` summarizing Score statistics and p-values separately
        for penalized (smooth) components of each term.
}
\description{
This function takes a fitted mgcv `gam` or `bam` object,
then performs Score test on the null (unpenalized/fixed) and penalized
(smooth) components of each smooth term.
It returns a clear summary as a `data.table`.
}
\examples{
\dontrun{
fit <- gam(y ~ s(x0) + s(x1), data=dat, method="REML")
taps_score_test(fit)
}

}
