% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taps_score_test.R
\name{taps_score_test}
\alias{taps_score_test}
\title{Perform Score Tests on mgcv Smooth Terms (Null and Penalized Components)}
\usage{
taps_score_test(fit, test.component = 1, null.tol = 1e-10)
}
\arguments{
\item{fit}{A fitted `gam` or `bam` model from the `mgcv` package.}

\item{test.component}{The index of smooth component to be tested.}

\item{null.tol}{The tolerance of row norm to detect indices of null space. Default to 1e-10.}
}
\value{
A `data.table` summarizing Score statistics and p-values separately
        for penalized (smooth) components of each term.
}
\description{
This function takes a fitted mgcv `gam` or `bam` object,
then performs Score test on the null (unpenalized/fixed) and penalized
(smooth) components of each smooth term.
It returns a clear summary as a `data.table`.
}
\examples{
\dontrun{
fit <- gam(y ~ s(x0) + s(x1), data=dat, method="REML")
taps_score_test(fit)
}

}
